ajout exemples + correction case

+++ src/compiler/primitives.rs
@@ -20,7 +20,7 @@ pub fn is_primitive(nom: &String) -> bool {
     PRIMITIVES.contains(&nom.as_str())
 }
 
-pub fn has_args(nom : &String, length:i32) -> i32 {
+pub fn has_args(nom : &String, length:usize) -> usize {
     if nom.clone().eq("not") {
         return length - 1;
     }
